Refactoring class name, user-facing messages
