+++ src/main/java/gateway/controller/AdminController.java
@@ -15,23 +15,21 @@
  **/
 package gateway.controller;
 
+import gateway.controller.util.PiazzaRestController;
+
 import java.util.HashMap;
 import java.util.Map;
+
+import model.response.ErrorResponse;
+import model.response.PiazzaResponse;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.HttpMediaTypeNotSupportedException;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
-import model.response.ErrorResponse;
-import model.response.PiazzaResponse;
 
 /**
  * REST Controller that defines administrative end points that reference
@@ -43,7 +41,7 @@ import model.response.PiazzaResponse;
  */
 @CrossOrigin
 @RestController
-public class AdminController extends ErrorController {
+public class AdminController extends PiazzaRestController {
 	@Value("${vcap.services.pz-kafka.credentials.host}")
 	private String KAFKA_ADDRESS;
 	@Value("${jobmanager.host}")
@@ -73,26 +71,26 @@ public class AdminController extends ErrorController {
 		// Return
 		return new ResponseEntity<Map<String, Object>>(stats, HttpStatus.OK);
 	}
-	
+
 	/**
 	 * Error handling for missing GET request
 	 * 
-	 * @return ResponseEntity<?> error describing the missing GET endpoint
+	 * @return Error describing the missing GET end point
 	 */
 	@RequestMapping(value = "/error", method = RequestMethod.GET)
 	public ResponseEntity<?> missingEndpoint_GetRequest() {
-		String message = "Missing GET request endpoint, please check the API.";
+		String message = "Gateway GET endpoint not defined. Please verify the API for the correct call.";
 		return new ResponseEntity<PiazzaResponse>(new ErrorResponse(null, message, "Gateway"), HttpStatus.BAD_REQUEST);
 	}
 
 	/**
 	 * Error handling for missing POST request
 	 * 
-	 * @return ResponseEntity<?> error describing the missing POST endpoint
+	 * @return Error describing the missing POST endpoint
 	 */
 	@RequestMapping(value = "/error", method = RequestMethod.POST)
 	public ResponseEntity<?> missingEndpoint_PostRequest() {
-		String message = "Missing POST request endpoint, please check the API.";
+		String message = "Gateway POST endpoint not defined. Please verify the API for the correct call.";
 		return new ResponseEntity<PiazzaResponse>(new ErrorResponse(null, message, "Gateway"), HttpStatus.BAD_REQUEST);
 	}
 }
