+++ src/main/java/gateway/controller/DeploymentController.java
@@ -16,6 +16,7 @@
 package gateway.controller;
 
 import gateway.controller.util.GatewayUtil;
+import gateway.controller.util.PiazzaRestController;
 
 import java.security.Principal;
 
@@ -50,7 +51,7 @@ import util.PiazzaLogger;
  */
 @CrossOrigin
 @RestController
-public class DeploymentController extends ErrorController {
+public class DeploymentController extends PiazzaRestController {
 	@Autowired
 	private GatewayUtil gatewayUtil;
 	@Autowired
