+++ src/main/java/gateway/controller/JobController.java
@@ -16,6 +16,7 @@
 package gateway.controller;
 
 import gateway.controller.util.GatewayUtil;
+import gateway.controller.util.PiazzaRestController;
 
 import java.security.Principal;
 
@@ -55,7 +56,7 @@ import util.PiazzaLogger;
  */
 @CrossOrigin
 @RestController
-public class JobController extends ErrorController {
+public class JobController extends PiazzaRestController {
 	@Autowired
 	private GatewayUtil gatewayUtil;
 	@Autowired
