+++ src/main/java/gateway/controller/util/PiazzaRestController.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  **/
-package gateway.controller;
+package gateway.controller.util;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -26,20 +26,32 @@ import model.response.ErrorResponse;
 import model.response.PiazzaResponse;
 
 /**
- * This rest controller manages controller errors, including missing endpoints 
- * and general spring exceptions which may occur during serialization/deserialization. 
+ * This controller manages controller errors, including missing endpoints and
+ * general spring exceptions which may occur during
+ * serialization/deserialization. Controllers will extend this class in order to
+ * get standardized error handling automatically injected into their endpoints.
  * 
  * @author Sonny.Saniev
  *
  */
 @RestController
-public class ErrorController {
-
+public class PiazzaRestController {
+	/**
+	 * Defines a generic error handler that Spring will use to form a REST
+	 * response to the user when some servlet processing exception has occurred.
+	 * This enables Piazza controllers to wrap generic Spring errors in our
+	 * PiazzaResponse interface; so that all responses from Piazza become
+	 * standardized.
+	 * 
+	 * @param exception
+	 *            The exception that has occurred
+	 * @return Appropriate Piazza Error response
+	 */
 	@ExceptionHandler(Exception.class)
 	@ResponseStatus(value = HttpStatus.BAD_REQUEST)
 	@ResponseBody
-	public ResponseEntity<PiazzaResponse> genericHandler(Exception e) {
-		String message = String.format("%s%s", "Piazza exception occurred. ", e.toString());
+	public ResponseEntity<PiazzaResponse> genericHandler(Exception exception) {
+		String message = String.format("Piazza Exception occurred: ", exception.getMessage());
 		return new ResponseEntity<PiazzaResponse>(new ErrorResponse(null, message, "Gateway"), HttpStatus.BAD_REQUEST);
 	}
 }
